Restore public accessors in QueueMessage files; add specific unit test; make connected tests more reliable
