+++ src/main/java/bio/terra/stairway/QueueMessageNoFields.java
@@ -4,14 +4,14 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 // We use this class to read the type information, to guide deserializing the right type
 @JsonIgnoreProperties(ignoreUnknown = true)
-class QueueMessageNoFields extends QueueMessage {
+public class QueueMessageNoFields extends QueueMessage {
   private QueueMessageType type;
 
-  QueueMessageType getType() {
+  public QueueMessageType getType() {
     return type;
   }
 
-  void setType(QueueMessageType type) {
+  public void setType(QueueMessageType type) {
     this.type = type;
   }
