+++ src/main/java/bio/terra/stairway/QueueMessageReady.java
@@ -4,7 +4,7 @@ import bio.terra.stairway.exception.DatabaseOperationException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-class QueueMessageReady extends QueueMessage {
+public class QueueMessageReady extends QueueMessage {
   private static final Logger logger = LoggerFactory.getLogger(QueueMessageReady.class);
 
   private QueueMessageType type;
@@ -12,14 +12,14 @@ class QueueMessageReady extends QueueMessage {
 
   private QueueMessageReady() {}
 
-  QueueMessageReady(String flightId) {
+  public QueueMessageReady(String flightId) {
     this.type =
         new QueueMessageType(QueueMessage.FORMAT_VERSION, QueueMessageEnum.QUEUE_MESSAGE_READY);
     this.flightId = flightId;
   }
 
   @Override
-  void process(Stairway stairway) throws InterruptedException {
+  public void process(Stairway stairway) throws InterruptedException {
     try {
       boolean resumed = stairway.resume(flightId);
       logger.info(
@@ -32,19 +32,19 @@ class QueueMessageReady extends QueueMessage {
     }
   }
 
-  QueueMessageType getType() {
+  public QueueMessageType getType() {
     return type;
   }
 
-  private void setType(QueueMessageType type) {
+  public void setType(QueueMessageType type) {
     this.type = type;
   }
 
-  String getFlightId() {
+  public String getFlightId() {
     return flightId;
   }
 
-  void setFlightId(String flightId) {
+  public void setFlightId(String flightId) {
     this.flightId = flightId;
   }
 }
